+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Includes only major changes (manual updates)
 # Version 0.8.2
 - added popup in pallet to select different line end (on current selected line)
+- added pallet entry to strip control chars
 
 # Version 0.8.1
 - improved mouse handle on workspace tab interaction